pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-legal-compliance-agent"
version = "0.1.0"
description = "웹사이트 법적 준수 여부를 멀티 에이전트로 분석하는 AI 시스템"
requires-python = ">=3.12"
dependencies = [
    # LLM / Embedding
    "openai>=1.35",
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langchain-community>=0.2",
    "langchain-text-splitters>=0.2",
    # Vector DB
    "qdrant-client>=1.9",
    "langchain-qdrant>=0.1",
    # Cache + Stream
    "redis>=5.0",
    # 법령 수집 / 파싱
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    # Token / Sparse Search
    "tiktoken>=0.7",
    "rank-bm25>=0.2",
    # API server
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    # UI
    "streamlit>=1.36",
    # Utilities
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "loguru>=0.7",
    "tqdm>=4.66",
    "apscheduler>=3.10",
    "networkx>=3.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "responses>=0.25",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
